+++ src/Spreads.Extensions/Persistence/SeriesInfo.cs
@@ -1,0 +1,45 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Spreads.Persistence {
+
+    /// <summary>
+    /// 
+    /// </summary>
+    public unsafe struct SeriesInfo {
+        // 
+        //public fixed byte SeriesId[16];
+
+        /// <summary>
+        /// If zero, the key type has variable size when serialized to binary, and the first 4 bytes in the binary representation are the length.
+        /// </summary>
+        public int KeyTypeSize;
+        public short KeyTypeCode;
+        /// <summary>
+        /// If zero, the value type has variable size when serialized to binary, and the first 4 bytes in the binary representation are the length.
+        /// </summary>
+        public int ValueTypeSize;
+        public short ValueTypeCode;
+    }
+
+    ///// <summary>
+    ///// Describes type properties needed for binary serialization
+    ///// </summary>
+    //public struct TypeInfo
+    //{
+    //    private int _size;
+
+    //    /// <summary>
+    //    /// If zero, the type has variable size when serialized to binary
+    //    /// </summary>
+    //    public int Size
+    //    {
+    //        get { return _size; }
+    //        set { _size = value; }
+    //    }
+    //}
+
+}